pyproject.toml
[tool.poetry]
name = "intent-networking"
version = "0.1.0"
description = "Intent Networking"
authors = ["Daniel Ashton <info@networktocode.com>"]
license = "Apache-2.0"
readme = "README.md"
homepage = "https://github.com/dashton956-alt/natuobot-iintent-network-app"
repository = "https://github.com/dashton956-alt/natuobot-iintent-network-app"
documentation = "https://docs.nautobot.com/projects/intent-networking/en/latest/"
keywords = ["nautobot", "nautobot-app", "nautobot-plugin"]
classifiers = [
    "Intended Audience :: Developers",
    "Development Status :: 5 - Production/Stable",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
packages = [
    { include = "intent_networking" },
]
include = [
    # Poetry by default will exclude files that are in .gitignore
    { path = "intent_networking/static/intent_networking/docs/**/*", format = ["sdist", "wheel"] }
]

[tool.poetry.dependencies]
python = ">=3.10,<3.14"
# Used for local development
nautobot = ">=3.0.0,<4.0.0"
requests = "^2.32.5"
jinja2 = "^3.1.6"
nornir = "^3.5.0"
nornir-netmiko = "^1.0.1"

[tool.poetry.group.dev.dependencies]
coverage = "*"
django-debug-toolbar = "*"
invoke = "*"
ipython = "*"
pylint = "*"
pylint-django = ">=2.5.4"
pylint-nautobot = ">=0.3.1"
ruff = "0.5.5"
yamllint = "*"
toml = "*"
# Python implementation of markdownlint
pymarkdownlnt = "~0.9.30"
Markdown = "*"
towncrier = ">=23.6.0,<=24.8.0"
to-json-schema = "*"
jsonschema = "*"
djlint = ">=1.36.4,<2.0.0"
djhtml = ">=3.0.8,<4.0.0"
tomli = { version = "*", markers = "python_version < '3.11'" }  # tomli is included in the stdlib as of Python 3.11

[tool.poetry.group.docs.dependencies]
# Render custom markdown for version added/changed/remove notes
markdown-version-annotations = "1.0.1"
# Rendering docs to HTML
mkdocs = "1.6.0"
# Material for MkDocs theme
mkdocs-material = "9.5.32"
# Automatic documentation from sources, for MkDocs
mkdocstrings = "0.25.2"
mkdocstrings-python = "1.10.8"
mkdocs-autorefs = "1.2.0"
griffe = "1.1.1"

[tool.poetry.extras]
all = [
]

[tool.pylint.master]
# Include the pylint_django plugin to avoid spurious warnings about Django patterns
load-plugins = "pylint_django, pylint_nautobot"
ignore = ".venv"

[tool.pylint.basic]
# No docstrings required for private methods (Pylint default), or for test_ functions, or for inner Meta classes.
no-docstring-rgx = "^(_|test_|Meta$)"

[tool.pylint.messages_control]
disable = """,
    line-too-long,
    too-many-positional-arguments,
    nb-warn-dunder-filter-field,
    nb-no-search-function,
    nb-no-char-filter-q,
    too-few-public-methods,
    too-many-ancestors,
    abstract-method,
    duplicate-code,
    import-error,
    import-outside-toplevel,
    nb-no-model-found,
    inconsistent-return-statements,
    invalid-name,
    too-many-locals,
    broad-exception-caught,
    arguments-differ,
"""

[tool.pylint.miscellaneous]
# Don't flag TODO as a failure, let us commit with things that still need to be done in the code
notes = """,
    FIXME,
    XXX,
    """

[tool.ruff]
line-length = 120
target-version = "py38"

[tool.ruff.lint]
select = [
    "D",  # pydocstyle
    "F", "E", "W",  # flake8
    "S",  # bandit
    "I",  # isort
]
ignore = [
    # warning: `one-blank-line-before-class` (D203) and `no-blank-line-before-class` (D211) are incompatible.
    "D203", # 1 blank line required before class docstring

    # D212 is enabled by default in google convention, and complains if we have a docstring like:
    # """
    # My docstring is on the line after the opening quotes instead of on the same line as them.
    # """
    # We've discussed and concluded that we consider this to be a valid style choice.
    "D212", # Multi-line docstring summary should start at the first line
    "D213", # Multi-line docstring summary should start at the second line

    # Produces a lot of issues in the current codebase.
    "D401", # First line of docstring should be in imperative mood
    "D407", # Missing dashed underline after section
    "D416", # Section name ends in colon
    "E501", # Line too long
]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.per-file-ignores]
"intent_networking/migrations/*" = [
    "D",
]
"intent_networking/tests/*" = [
    "D",
    "S"
]

[tool.coverage.run]
disable_warnings = ["already-imported"]
relative_files = true
omit = [
    # Skip Tests
    "*/tests/*",
]
include = [
    "intent_networking/*",
]

[tool.pymarkdown]
plugins.selectively_enable_rules = true
plugins.blanks-around-lists.enabled = true
plugins.list-anchored-indent.enabled = true
extensions.front-matter.enabled = true

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.towncrier]
package = "intent_networking"
directory = "changes"
filename = "docs/admin/release_notes/version_X.Y.md"
template = "development/towncrier_template.j2"
start_string = "<!-- towncrier release notes start -->"
issue_format = "[#{issue}](https://github.com/dashton956-alt/natuobot-iintent-network-app/issues/{issue})"

[[tool.towncrier.type]]
directory = "breaking"
name = "Breaking Changes"
showcontent = true

[[tool.towncrier.type]]
directory = "security"
name = "Security"
showcontent = true

[[tool.towncrier.type]]
directory = "added"
name = "Added"
showcontent = true

[[tool.towncrier.type]]
directory = "changed"
name = "Changed"
showcontent = true

[[tool.towncrier.type]]
directory = "deprecated"
name = "Deprecated"
showcontent = true

[[tool.towncrier.type]]
directory = "removed"
name = "Removed"
showcontent = true

[[tool.towncrier.type]]
directory = "fixed"
name = "Fixed"
showcontent = true

[[tool.towncrier.type]]
directory = "dependencies"
name = "Dependencies"
showcontent = true

[[tool.towncrier.type]]
directory = "documentation"
name = "Documentation"
showcontent = true

[[tool.towncrier.type]]
directory = "housekeeping"
name = "Housekeeping"
showcontent = true

[tool.djlint]
ignore = "H005,H006,H008,H011,H012,H013,H014,H015,H016,H019,H020,H021,H022,H023,H025,H026,H030,H031,H037"
profile="django"
extend_exclude = "docs/,intent_networking/static/"
